Sat, Mar 19, 2022, 12:23 AM +00:00

+++ extra/extra4.c
@@ -22,7 +22,7 @@ int main(void)
         sum = sum + scores[i];
     }
 
-    float average = sum / n;
+    float average = (float) sum / (float) n;
 
     //printing average
     printf("%f\n", average);
